Update to general store
Created more streamlined selling process as well as made it so when you sell items you get coins in return.

you no longer have the option to sell 1 of a singular item.

changed general store 'help menu'

you can no type "sell bag" and wait 10 seconds to sell your entire bag
note for this: i may add a simple algorithm to check if you have over a certain amount of items in your bag and that will correlate to the amount of time you have to wait. as well as adding a 'are you sure' promt  before having to wait the proposed amount of time.